weather_api/requirements.txt
requests
mysql-connector-python
python-decouple